chore: operationPage component to fetch data for grid
